requirements.txt
anyio==4.7.0
asgiref==3.8.1
beautifulsoup4==4.12.3
blinker==1.4
certifi==2024.7.4
charset-normalizer==3.3.2
cryptography==3.4.8
Cython==0.29.34
dbus-python==1.2.18
distro==1.7.0
Django==5.1.4
dnspython==2.6.1
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.7
httplib2==0.20.2
httpx==0.28.1
idna==3.7
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
meson==0.61.2
more-itertools==8.10.0
netifaces==0.11.0
numpy==2.0.1
oauthlib==3.2.0
pandas==2.2.2
pycairo==1.20.1
PyGObject==3.42.1
PyJWT==2.3.0
pymongo==4.8.0
pyparsing==2.4.7
pyTelegramBotAPI==4.25.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.9
pytz==2024.1
PyYAML==5.3.1
requests==2.32.3
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.5.3
ssh-import-id==5.11
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
wadllib==1.3.6
zipp==1.0.0
